feat: 创建认证和数据集相关表结构，加载初始数据

+++ avalon-modules/avalon-authorization/src/main/java/io/github/lishangbu/avalon/authorization/repository/UserRepository.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface UserRepository
-    extends ListCrudRepository<User, Long>, ListPagingAndSortingRepository<User, Long> {
+    extends ListCrudRepository<User, Integer>, ListPagingAndSortingRepository<User, Integer> {
   /**
    * 根据用户名查询用户信息
    *
